+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 - Add accuracy in results when interpolate and sort by it
 
 # Version Fix 1.0.1
+- Import proj_lat/lon left and right of polygon
+- Compute housenumber of proj_lat/lon left and right of polygon
+- Import it into database and use for compute vertex, interpolation